+++ apps/agent/cmd/agent/agent.go
@@ -5,18 +5,25 @@ import (
 	"fmt"
 	"os"
 	"os/signal"
+	"runtime/debug"
+	"strings"
 	"syscall"
 	"time"
 
+	"github.com/unkeyed/unkey/apps/agent/pkg/cluster"
 	"github.com/unkeyed/unkey/apps/agent/pkg/config"
 	"github.com/unkeyed/unkey/apps/agent/pkg/connect"
 	"github.com/unkeyed/unkey/apps/agent/pkg/load"
+	"github.com/unkeyed/unkey/apps/agent/pkg/membership"
 	"github.com/unkeyed/unkey/apps/agent/pkg/metrics"
-	"github.com/unkeyed/unkey/apps/agent/pkg/services/ratelimit"
 	"github.com/unkeyed/unkey/apps/agent/pkg/tinybird"
 	"github.com/unkeyed/unkey/apps/agent/pkg/tracing"
 	"github.com/unkeyed/unkey/apps/agent/pkg/uid"
+	"github.com/unkeyed/unkey/apps/agent/pkg/version"
 	"github.com/unkeyed/unkey/apps/agent/services/eventrouter"
+	"github.com/unkeyed/unkey/apps/agent/services/ratelimit"
+	"github.com/unkeyed/unkey/apps/agent/services/vault"
+	"github.com/unkeyed/unkey/apps/agent/services/vault/storage"
 	"github.com/urfave/cli/v2"
 )
 
@@ -43,21 +50,32 @@ func run(c *cli.Context) error {
 	if err != nil {
 		return err
 	}
+
+	if cfg.NodeId == "" {
+		cfg.NodeId = uid.Node()
+
+	}
+	if cfg.Region == "" {
+		cfg.Region = "unknown"
+	}
 	logger, err := setupLogging(cfg)
 	if err != nil {
 		return err
 	}
-	if cfg.NodeId == "" {
-		cfg.NodeId = uid.Node()
-	}
-	logger = logger.With().Str("nodeId", cfg.NodeId).Str("region", cfg.Region).Logger()
+	logger = logger.With().Str("nodeId", cfg.NodeId).Str("platform", cfg.Platform).Str("region", cfg.Region).Str("version", version.Version).Logger()
+	logger.Info().Strs("env", os.Environ()).Send()
+	// Catch any panics now after we have a logger but before we start the server
+	defer func() {
+		if r := recover(); r != nil {
+			logger.Panic().Interface("panic", r).Bytes("stack", debug.Stack()).Msg("panic")
+		}
+	}()
 
-	logger.Info().Str("file", configFile).Interface("cfg", cfg).Msg("configuration loaded")
+	logger.Info().Str("file", configFile).Msg("configuration loaded")
 
-	tracer := tracing.NewNoop()
 	{
-		if cfg.Tracing != nil {
-			t, closeTracer, err := tracing.New(context.Background(), tracing.Config{
+		if cfg.Tracing != nil && cfg.Tracing.Axiom != nil {
+			closeTracer, err := tracing.Init(context.Background(), tracing.Config{
 				Dataset:     cfg.Tracing.Axiom.Dataset,
 				Application: "agent",
 				Version:     "1.0.0",
@@ -72,7 +90,6 @@ func run(c *cli.Context) error {
 					logger.Error().Err(err).Msg("failed to close tracer")
 				}
 			}()
-			tracer = t
 			logger.Info().Msg("tracing to axiom")
 		}
 	}
@@ -108,30 +125,39 @@ func run(c *cli.Context) error {
 		}
 	}
 
-	srv, err := connect.New(connect.Config{Logger: logger, Tracer: tracer})
+	srv, err := connect.New(connect.Config{Logger: logger, Image: cfg.Image})
 	if err != nil {
 		return err
 	}
 
-	if cfg.Services.Ratelimit != nil {
-		rl, err := ratelimit.New(ratelimit.Config{
-			Logger: logger,
+	if cfg.Services.Vault != nil {
+		s3, err := storage.NewS3(storage.S3Config{
+			S3URL:             cfg.Services.Vault.S3Url,
+			S3Bucket:          cfg.Services.Vault.S3Bucket,
+			S3AccessKeyId:     cfg.Services.Vault.S3AccessKeyId,
+			S3AccessKeySecret: cfg.Services.Vault.S3AccessKeySecret,
+			Logger:            logger,
 		})
 		if err != nil {
-			logger.Fatal().Err(err).Msg("failed to create service")
+			return fmt.Errorf("failed to create s3 storage: %w", err)
 		}
-		rl = ratelimit.WithTracing(tracer)(rl)
+		v, err := vault.New(vault.Config{
+			Logger:     logger,
+			Storage:    s3,
+			Metrics:    m,
+			MasterKeys: strings.Split(cfg.Services.Vault.MasterKeys, ","),
+		})
+		if err != nil {
+			return fmt.Errorf("failed to create vault: %w", err)
+		}
 
-		rlServer := connect.NewRatelimitServer(rl, logger)
-
-		srv.AddService(rlServer)
-		logger.Info().Msg("started ratelimit service")
+		srv.AddService(connect.NewVaultServer(v, logger))
+		logger.Info().Msg("started vault service")
 	}
 
 	if cfg.Services.EventRouter != nil {
 		er, err := eventrouter.New(eventrouter.Config{
 			Logger:        logger,
-			Tracer:        tracer,
 			BatchSize:     cfg.Services.EventRouter.Tinybird.BatchSize,
 			BufferSize:    cfg.Services.EventRouter.Tinybird.BufferSize,
 			FlushInterval: time.Duration(cfg.Services.EventRouter.Tinybird.FlushInterval) * time.Second,
@@ -143,7 +169,60 @@ func run(c *cli.Context) error {
 		srv.AddService(er)
 	}
 
-	err = srv.Listen(fmt.Sprintf(":%d", cfg.Port))
+	var clus cluster.Cluster
+	if cfg.Cluster != nil {
+		memb, err := membership.New(membership.Config{
+
+			NodeId:   cfg.NodeId,
+			RpcAddr:  cfg.Cluster.RpcAddr,
+			Logger:   logger,
+			RedisUrl: cfg.Cluster.RedisUrl,
+		})
+		if err != nil {
+			return fmt.Errorf("failed to create membership: %w", err)
+		}
+
+		clus, err = cluster.New(cluster.Config{
+			NodeId:     cfg.NodeId,
+			RpcAddr:    cfg.Cluster.RpcAddr,
+			Membership: memb,
+			Logger:     logger,
+			Debug:      true,
+			AuthToken:  cfg.Cluster.AuthToken,
+		})
+		if err != nil {
+			return fmt.Errorf("failed to create cluster: %w", err)
+		}
+		defer func() {
+			err := clus.Shutdown()
+			if err != nil {
+				logger.Error().Err(err).Msg("failed to shutdown cluster")
+			}
+		}()
+
+		_, err = clus.Join([]string{})
+		if err != nil {
+			return fmt.Errorf("failed to join cluster: %w", err)
+		}
+
+		srv.AddService(connect.NewClusterServer(clus, logger))
+	}
+
+	if cfg.Services.Ratelimit != nil {
+		rl, err := ratelimit.New(ratelimit.Config{
+			Logger:  logger,
+			Cluster: clus,
+		})
+		if err != nil {
+			logger.Fatal().Err(err).Msg("failed to create service")
+		}
+		rl = ratelimit.WithTracing(rl)
+
+		srv.AddService(connect.NewRatelimitServer(rl, logger))
+		logger.Info().Msg("started ratelimit service")
+	}
+
+	err = srv.Listen(fmt.Sprintf(":%s", cfg.Port))
 	if err != nil {
 		return err
 	}
@@ -157,8 +236,10 @@ func run(c *cli.Context) error {
 }
 
 type configuration struct {
-	NodeId  string `json:"nodeId,omitempty" description:"A unique node id"`
-	Logging *struct {
+	Platform string `json:"platform,omitempty" description:"The platform this agent is running on"`
+	NodeId   string `json:"nodeId,omitempty" description:"A unique node id"`
+	Image    string `json:"image,omitempty" description:"The image this agent is running"`
+	Logging  *struct {
 		Axiom *struct {
 			Dataset string `json:"dataset" minLength:"1" description:"The dataset to send logs to"`
 			Token   string `json:"token" minLength:"1" description:"The token to use for authentication"`
@@ -181,7 +262,7 @@ type configuration struct {
 
 	Schema    string `json:"$schema,omitempty" description:"Make jsonschema happy"`
 	Region    string `json:"region,omitempty" description:"The region this agent is running in"`
-	Port      int    `json:"port,omitempty" max:"65535" min:"0" default:"8080" description:"Port to listen on"`
+	Port      string `json:"port,omitempty" default:"8080" description:"Port to listen on"`
 	Heartbeat *struct {
 		URL      string `json:"url" minLength:"1" description:"URL to send heartbeat to"`
 		Interval int    `json:"interval" min:"1" description:"Interval in seconds to send heartbeat"`
@@ -200,7 +281,22 @@ type configuration struct {
 				BatchSize     int    `json:"batchSize" min:"1" description:"Size of the batch"`
 			} `json:"tinybird,omitempty" description:"Send events to tinybird"`
 		} `json:"eventRouter,omitempty" description:"Route events"`
+		Vault *struct {
+			S3Bucket          string `json:"s3Bucket" minLength:"1" description:"The bucket to store secrets in"`
+			S3Url             string `json:"s3Url" minLength:"1" description:"The url to store secrets in"`
+			S3AccessKeyId     string `json:"s3AccessKeyId" minLength:"1" description:"The access key id to use for s3"`
+			S3AccessKeySecret string `json:"s3AccessKeySecret" minLength:"1" description:"The access key secret to use for s3"`
+			MasterKeys        string `json:"masterKeys" minLength:"1" description:"The master keys to use for encryption, comma separated"`
+		} `json:"vault,omitempty" description:"Store secrets"`
 	} `json:"services"`
+
+	Cluster *struct {
+		AuthToken string `json:"authToken" minLength:"1" description:"The token to use for http authentication"`
+		// SerfAddr string `json:"serfAddr" minLength:"1" description:"The address to use for serf"`
+		RedisUrl string `json:"redisUrl" minLength:"1" description:"The url to use for redis"`
+		RpcAddr  string `json:"rpcAddr" minLength:"1" description:"This node's internal address, including protocol and port"`
+		// Join     string `json:"join,omitempty"  description:"Addresses to join, comma separated"`
+	} `json:"cluster,omitempty"`
 }
 
 // TODO: generating this every time is a bit stupid, we should make this its own command