+++ apps/agent/pkg/cluster/node.go
@@ -1,0 +1,6 @@
+package cluster
+
+type Node struct {
+	Id      string
+	RpcAddr string
+}