+++ apps/agent/pkg/connect/service.go
@@ -2,16 +2,14 @@ package connect
 
 import (
 	"context"
-	"fmt"
+	"encoding/json"
 	"sync"
-	"time"
 
 	"net/http"
 
 	"github.com/bufbuild/connect-go"
 	ratelimitv1 "github.com/unkeyed/unkey/apps/agent/gen/proto/ratelimit/v1"
 	"github.com/unkeyed/unkey/apps/agent/pkg/logging"
-	"github.com/unkeyed/unkey/apps/agent/pkg/tracing"
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
 )
@@ -23,26 +21,26 @@ type Service interface {
 type Server struct {
 	sync.Mutex
 	logger         logging.Logger
-	tracer         tracing.Tracer
 	mux            *http.ServeMux
 	shutdownC      chan struct{}
 	isShuttingDown bool
 	isListening    bool
+	image          string
 }
 
 type Config struct {
 	Logger logging.Logger
-	Tracer tracing.Tracer
+	Image  string
 }
 
 func New(cfg Config) (*Server, error) {
 
 	return &Server{
 		logger:         cfg.Logger,
-		tracer:         cfg.Tracer,
 		isListening:    false,
 		isShuttingDown: false,
 		mux:            http.NewServeMux(),
+		image:          cfg.Image,
 
 		shutdownC: make(chan struct{}),
 	}, nil
@@ -52,7 +50,7 @@ func (s *Server) AddService(svc Service) {
 	pattern, handler := svc.CreateHandler()
 	s.logger.Info().Str("pattern", pattern).Msg("adding service")
 
-	h := newTracingMiddleware(newHeaderMiddleware(handler), s.tracer)
+	h := newTracingMiddleware(newHeaderMiddleware(newLoggingMiddleware(handler, s.logger)))
 	s.mux.Handle(pattern, h)
 }
 
@@ -73,8 +71,15 @@ func (s *Server) Listen(addr string) error {
 	s.Unlock()
 
 	s.mux.HandleFunc("/v1/liveness", func(w http.ResponseWriter, r *http.Request) {
+
+		b, err := json.Marshal(map[string]string{"status": "serving", "image": s.image})
+		if err != nil {
+			s.logger.Error().Err(err).Msg("failed to marshal response")
+			return
+		}
+
 		w.WriteHeader(http.StatusOK)
-		_, err := w.Write([]byte(fmt.Sprintf("OK, %s", time.Now().String())))
+		_, err = w.Write(b)
 		if err != nil {
 			s.logger.Error().Err(err).Msg("failed to write response")
 		}