+++ apps/agent/services/eventrouter/service.go
@@ -26,12 +26,10 @@ type Config struct {
 
 	Tinybird *tinybird.Client
 	Logger   logging.Logger
-	Tracer   tracing.Tracer
 }
 
 type service struct {
 	logger  logging.Logger
-	tracer  tracing.Tracer
 	batcher batch.BatchProcessor[event]
 	tb      *tinybird.Client
 }
@@ -50,12 +48,11 @@ func New(config Config) (*service, error) {
 				eventsByDatasource[e.datasource] = []any{}
 			}
 			eventsByDatasource[e.datasource] = append(eventsByDatasource[e.datasource], e.row)
-
 		}
 		for datasource, rows := range eventsByDatasource {
 			err := config.Tinybird.Ingest(datasource, rows)
 			if err != nil {
-				config.Logger.Err(err).Str("datasource", datasource).Int("rows", len(rows)).Msg("Error ingesting")
+				config.Logger.Err(err).Str("datasource", datasource).Interface("rows", rows).Msg("Error ingesting")
 			}
 			config.Logger.Info().Str("datasource", datasource).Int("rows", len(rows)).Msg("Ingested")
 		}
@@ -70,14 +67,17 @@ func New(config Config) (*service, error) {
 	return &service{
 		logger:  config.Logger,
 		batcher: *batcher,
-		tracer:  config.Tracer,
 		tb:      config.Tinybird,
 	}, nil
 }
 
 func (s *service) CreateHandler() (string, http.Handler) {
 	return "/v0/events", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx, span := s.tracer.Start(r.Context(), "events")
+		start := time.Now()
+		defer func() {
+			s.logger.Info().Str("method", r.Method).Str("path", r.URL.Path).Int64("serviceLatency", time.Since(start).Milliseconds()).Msg("request")
+		}()
+		ctx, span := tracing.Start(r.Context(), tracing.NewSpanName("eventrouter", "v0/events"))
 		defer span.End()
 
 		err := auth.Authorize(ctx, r.Header.Get("Authorization"))