+++ apps/api/CHANGELOG.md
@@ -1,5 +1,26 @@
 # api
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [9b80f19]
+  - @unkey/cache@1.0.2
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [a2578ed]
+  - @unkey/cache@1.0.1
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [1262b44]
+  - @unkey/cache@1.0.0
+
 ## 0.0.5
 
 ### Patch Changes